Added GeofenceTransitionIntentService, this currently contains a lot of google's example code that is commented out to use later.
Added a save method in the geofencerepository

+++ src/main/java/com/hgyllensvard/geofencemanager/geofence/persistence/GeofenceModel.java
@@ -9,11 +9,7 @@ public class GeofenceModel extends RealmObject {
     @PrimaryKey
     public String name;
 
-    public long latitude;
+    public double latitude;
 
-    public long longitude;
-
-
-
-
+    public double longitude;
 }
